much simpler android flags for streaming (graphviz is only major copied code)
