+++ .gitignore
@@ -85,3 +85,7 @@ etc/FFmpeg
 
 # qemu core dumps
 web/cmd/server/*.core
+
+# gomobile / android
+*.aar
+*.jar
