+++ robot/web/graphviz_nocgo.go
@@ -1,0 +1,10 @@
+//go:build !no_cgo || android
+
+package web
+
+import "net/http"
+
+// stub for missing graphviz
+func (svc *webService) handleVisualizeResourceGraph(w http.ResponseWriter, r *http.Request) {
+	w.Write([]byte(`<html><body>Resource graph not supported on this build</body></html>`))
+}
