+++ robot/web/web_notc.go
@@ -1,4 +1,4 @@
-//go:build no_cgo
+//go:build no_cgo && !android
 
 package web
 
@@ -70,6 +70,3 @@ func (svc *webService) initStreamServer(ctx context.Context, options *weboptions
 
 // stub for missing gostream
 type options struct{}
-
-// stub for missing graphviz
-func (svc *webService) handleVisualizeResourceGraph(w http.ResponseWriter, r *http.Request) {}
