+++ web/server/entrypoint_cgo.go
@@ -1,4 +1,4 @@
-//go:build !no_cgo
+//go:build !no_cgo || android
 
 package server
